Removes destructor tests
The destructor tests in there current form will always fail. This is because the if check is always true since calling delete only deletes the memory allocated on the heap, and does not convert the pointer (which is allocated on the stack) to a nullpointer, which the tests as written are assuming. 

I don't believe testing destructors is necessary anyways, which just need to check the methods, so I think we can just delete them

also fyi the syntax for calling delete was incorrect